lead-generation-api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and authentication
supabase>=2.3.0
PyJWT>=2.8.0

# AI services
openai>=1.3.0
anthropic>=0.7.0

# Apify client (for n8n workflow replication)
apify-client>=1.7.1

# HTTP client - let pip resolve the best version
httpx

# Async support
asyncio-throttle==1.0.2

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.0

# Optional: Lead enrichment services
# apollo-python==1.0.0  # Uncomment if using Apollo
# hunter-python==1.0.0  # Uncomment if using Hunter
# clearbit-python==1.0.0  # Uncomment if using Clearbit